% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{as_crs}
\alias{as_crs}
\title{Create a CRS from an object}
\usage{
as_crs(x, ...)
}
\arguments{
\item{x}{An object that can be interpreted as a CRS}

\item{...}{Unused}
}
\value{
An object of (at least) class
}
\description{
Create a CRS from an object
}
\examples{
as_crs(4326)
as_crs("+proj=longlat +datum=WGS84 +no_defs +type=crs")
as_crs('
  GEOGCS[
    "WGS 84",
    DATUM[
      "WGS_1984",
      SPHEROID["WGS 84",6378137,298.257223563, AUTHORITY["EPSG","7030"]],
      AUTHORITY["EPSG","6326"]
    ],
    PRIMEM["Greenwich",0, AUTHORITY["EPSG","8901"]],
    UNIT["degree",0.0174532925199433, AUTHORITY["EPSG","9122"]],
    AUTHORITY["EPSG","4326"]
   ]
')

as_crs('
{
  "type":"GeographicCRS",
  "name":"WGS 84",
  "datum":{
    "type":"GeodeticReferenceFrame",
    "name":"World Geodetic System 1984",
    "ellipsoid":{
      "name":"WGS 84",
      "semi_major_axis":6378137,
      "inverse_flattening":298.257223563
     }
   },
   "coordinate_system":{
     "subtype":"ellipsoidal",
     "axis":[
       {
         "name":"Geodetic latitude",
         "abbreviation":"Lat",
         "direction":"north","unit":"degree"
       },
       {
         "name":"Geodetic longitude",
         "abbreviation":"Lon",
         "direction":"east","unit":"degree"
       }
     ]
   },
   "area":"World",
   "bbox":{
     "south_latitude":-90,
     "west_longitude":-180,
     "north_latitude":90,
     "east_longitude":180
   },
   "id":{
   "authority":"EPSG",
   "code":4326
  }
}
')

}
