% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocrs-package.R
\docType{package}
\name{geocrs-package}
\alias{geocrs}
\alias{geocrs-package}
\title{geocrs: Create and Validate Coordinate Reference System Definitions}
\description{
Creates and validates Coordinate Reference System (CRS) definitions.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
