% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geocrs_crs_view}
\alias{geocrs_crs_view}
\alias{geocrs_crs_area_of_use}
\alias{geocrs_crs_json}
\alias{geocrs_crs_proj4}
\alias{geocrs_crs_wkt2}
\alias{geocrs_crs_wkt_gdal}
\title{PROJ database}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9892 rows and 10 columns.}
\source{
The \href{https://proj.org/}{PROJ} sources
}
\usage{
geocrs_crs_view

geocrs_crs_area_of_use

geocrs_crs_json

geocrs_crs_proj4

geocrs_crs_wkt2

geocrs_crs_wkt_gdal
}
\description{
Partial cache of the PROJ database.
}
\keyword{datasets}
